Fixed a bug in the savestate system that caused memcards to eject when they shouldn't have (had uint instead of u64 >_<).
Developers: Changed the way PCSX2_ALIGNED16 macros work, so that they're more friendly to MSVC and Visual Assist X intellisense (more more red squigglies on vars like cpuRegs!)

git-svn-id: http://pcsx2.googlecode.com/svn/trunk@618 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ pcsx2/Sio.cpp
@@ -599,7 +599,7 @@ void SaveState::sioFreeze()
 	
 		for( int i=0; i<2; ++i )
 		{
-			uint newCRC = MemoryCard::GetCRC( i );
+			u64 newCRC = MemoryCard::GetCRC( i );
 			if( newCRC != m_mcdCRCs[i] )
 			{
 				m_PostSavestateCards[i] = 64;
